Split general buttons and update animated icon button
